refactored electron index
This is the beginning of the refactoring effort
