+++ cspell.json
@@ -1,1 +1,1 @@
-{"version":"0.2","flagWords":[],"language":"en","words":["nextui"]}
+{"flagWords":[],"words":["nextui","linebreak"],"language":"en","version":"0.2"}
