+++ package.json
@@ -18,9 +18,9 @@
   },
   "homepage": "https://github.com/select-studios/Select-Launcher#readme",
   "devDependencies": {
-    "concurrently": "^7.6.0"
-  },
-  "dependencies": {
-    "eslint": "^8.40.0"
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
+    "concurrently": "^7.6.0",
+    "eslint": "^8.52.0"
   }
 }
